core/vm: add new authcall evm method
